Only override 'guitablabel' if user hasn't already set it
If guitablabel is already set in .vimrc, don't have the MacVim default
gvimrc override it. Even though Vim's docs recommend putting GUI
settings in gvimrc, that's mostly for technical reasons, and MacVim
should still aim to respect settings that a user has already set. This
also makes it behave more similarly with Vim (you can set 'guitablabel'
in vimrc or gvimrc and it would still work).

This was previously done in #635 but reverted, but #899 raised this
issue again.

Close #899