Make MMTitlebarAppearsTransparent option work on pre-Mojave
Added the special path to make tranparent titlebars work on pre-Mojave
builds as well. It's annoying we have to two maintain separate code
paths for how windows / titlebars work but this wasn't too hard to add.
We just need to make sure to to set the background color on the window
*only* when we have transparent titlebar set. Otherwise we don't want to
do it because of weird interactions with textured window flag.

This addresses comments made in #888.