Fix drag-and-drop with swap file enabled causing MacVim to freeze
MMBackend's `handleOpenWithArguments` previously set a "flushDisabled"
flag to disable all flushing during handling the files. This was unsafe,
because under the new code that handles editing immediately (instead of
building a deferred set of Ex commands) MacVim tries to immediately pop
up a dialog box but that message doesn't get properly flushed because
it's disabled. Just remove the setting of that disabled flag as it
doesn't seem like it's gaining anything.

If we want to have a way to disable flushing for performance reasons in
the future , we should make sure the "force" flag in flushQueue:
actually gets respected and used properly (only when we want to force
it). Right now the "force" flag isn't actually used.

Also, make sure the dialog handling code handles the "no GUI resize"
resize message (which gets set when we have `guioptions+=k`) as well and
don't drop the message. Otherwise if drag-and-drop opens a new tab
(since there are multiple files dropped) and the user has guioptions+=k,
it won't get redrawn properly because the
SetTextDimensionsNoResizeWindowMsgID message would get dropped. It's
unfortunate it's a hardcoded hack like this and this should be revisited
in the future.

Fix #913