Fix flaky test_suspend script test
Make sure to wait for shell prompt to come up when quitting Vim before
sending the "exit" command to shell. This seems to help in certain
situations where it appears to have a race condition where
`term_wait(buf)` isn't enough to wait for Vim to have fully quitted and
returning contorl back to the shell.