Revert "Enable hardened runtime"
The hardened runtime was causing issues with misc unsigned frameworks
and failing CI, and also making it hard to locally debug code. Because
we currently perform code signing outside of Xcode anyway by using the
`codesign` utility, it's not particularly useful to embed hardened
runtime in the project file itself. We just need to sign the app with
`-o runtime` flag when signing it before sending over to notarization
service.

This reverts commit 76cb94be7d976e5e5fb15e6018881aad8aec2967.