Switch Travis CI release to use Github release drafts
This is necessary because with the new macOS requirements to use code
signing and app notarization, there's a set of additional steps after
Travis CI has built the app to be done. Using drafts would upload the
built app to a non-release location to give time to finish the remaining
signing and notarization steps without creating an incomplete Github
release.