Use a custom I-beam mouse cursor for pre-Mojave OS versions
Prevoiusly, #859 removed the custom high-contrast I-beam mouse cursor in
favor of using the system-default I-beam cursor. However, it mostly
worked well on 10.14 Mojave or above, as the cursor is designed with
dark mode in mind and works well on both bright and dark backgrounds.
Since this doesn't work well on older macOS versions, add back the
custom cursor and only enable it on those older macOS versions.

Fix #910.