Make Sparkle updater an optional component of MacVim
MacVim by default uses Sparkle to update itself. It checks a URL that
tells it whether there's a new binary release, and if one exists, it
will download it and patches MacVim. However, there are certain
situations where this is not wanted (e.g. when MacVim is bundled as part
of a package management system), and previously there was no way to
disable Sparkle.

Make Sparkle optional by implementing a new configure argument:
`./configure --disable-sparkle`, which will set a compile-time ifdef.
Instead of instantiating the updater in MainMenu.nib as suggested by
Sparkle's documentation, just instantiate it manually in code, and also
disable the "Check for updates" menu item and updater-related preference
pane items to make sure these settings don't confuse users when they are
running a Sparkle-disabled build.

Sparkle.framework is currently still bundled when it's disabled, because
there doesn't seem to be an easy way to configure the copy build phase.
It could be manually removed in a script after the build though, as it's
now optionally linked to instead of mandatory.

Fix #849.