Add new appearance option for hiding title bar
This adds a new appearance option in Preferences to pick hiding title
bar (in addition to using the transparent look). It was already an
option before but this exposes it in the UI. Also, fix setting title to
cache properly.

Also fix non-native fullscreen to respect appearance changing in the
Preferences pane.

Also, see #1038 which added appareance pane.