Fix mvim:// protocol handler not working well with iTerm2
iTerm2's MacVim integration generates URLs that has the file separator
("/") escaped as ("%2F"). This is incorrect (as it implies the slash is
part of the file name itself instead of a separator), but work around it
for now. Instead of using NSURL to parse the file:// URL, just decode
the input parameter and manually parse it. This still handles the
special characters like space or "?", but will handle the escaped
slashes as well. Given that "/" is not a valid filename character we
should not run into ambiguity here.

Also, add dialog boxes to show an error if MacVim doesn't know how to
handle a file path.

Will add regression tests for these edge cases in a future commit.

Fix #1020