Fix disk image build target (macvim-dmg) permission issues on Travis
create-dmg uses osascript calls to prettify the created dmg file, which
has not been working in Travis CI for a while since it would pop up a
dialog box asking for permission in the later macOS versions, and
there's no way to interact with it in CI. The old version of create-dmg
silently handles the failure but the new version just fails the build,
so we need to manually pass in --skip-jenkins flag to skip the prettify
step.

This is still ok, because currently the maintainer still has to manually
re-build and re-sign the dmg file locally after Travis built it to add
signing and notarization to the app, so the dmg file built by Travis is
just an intermediate step, and not going to be used for the final
release.