Fix switching to fullscreen loses keyboard focus if titlebar is hidden
For some reason, when the titlebar is hidden, if we use `set fullscreen`
in native full screen mode, the focus will be lost and the user needs to
manually click on the screen to be able to type in MacVim. Current
theory is that because fullscreen window always has a title bar,
sometime that resulted in the window being defocused when the OS is
fixing it up.

For an easy fix, simply assert the focus by using `makeFirstResponder`
on the text view. This may not work well if the program is more
complicated but for MacVim the text view should always be where the
focus goes.

Fix #1078