Switch download link to directly link to dmg file
Also, provide a separate link to link to the release notes. This is more
in tune with what other software projects do and more user-friendly as
most people just want a direct link to download the dmg file.

This does create more release burden as we now have to go through and
update the links every time we release but it's not a big deal.