Help menu search will now search Vim documentation
Implement macOS Help menu's search functionality, so that it can be used
to search Vim's documentation. For now, can use space-delimited search
string to search Vim's doc tags. The search results will display
something like "options.txt > 'termwinsize'" when searching for "term
size".

Currently this only works with Vim's built-in documentation. Due to the
asynchronous nature of the search, it's a little tricky to support
plugins as different Vim instances could be loading different plugins.
For now, just the built-in Vim documentation should serve most of the
needs.

Also, properly set the help menu on the app so that localized menus will
still show the search box (previously it had to be called exactly
'Help').