Fix Xcode project file to only require Xcode 6.3
Previously the objectVersion in the pbxproj file was erroneously changed
from 46 to 54 (in fcd579756) as the Xcode editor automatically changed
that. Fix this by manually change it back to an older version, but
change it to 47 instead as that covers all the way up to Xcode 6.3 which
is definitely old enough to be backward compatible. In the future will
need to look into ways to lock that in place to prevent accidental
changes.

Fix #1100