Decouple Sparkle from MMAppController if DISABLE_SPARKLE is set
This makes the code buildable if DISABLE_SPARKLE is set, even if the
Sparkle framework is removed. This doesn't mean there is a way to build
without linking against Sparkle though as Xcode doesn't provide an easy
way to control whether a framework is used or not via xcodebuild command
line arguemnts, but for situations like Apple Silicon builds we can at
least just remove the reference to Sparkle and it should build now.