Fix WinBar menus in GUI, and off-by-one error when adding new menus
Fix WinBar menus to not create menu items in the main menu bar.

Also, fix an off-by-one error when adding new menu items to the main
menu when adding the menus using priorities to insert them in the
middle. This also allowed adding a menu (when using index 0) that goes
before the main "MacVim" app menu which is quite wrong as the app menu
should always be the first one.

Fix #918