Make SF Symbols double-width
macOS 11 (Big Sur) has added SF Symbols support for the OS. While these
are primarily designed to be used in apps as images, each symbol does
indeed correspond to a Unicode entry and can be used alongside normal
texts in a text editor, similar to how emojis work. Since they tend to
be pretty square in shape, we should treat them as double-width like
emojis.

The Unicode range was extracted by using Apple's SF Symbols app, and
copying out all the symbols and inspecting them. It will likely need to
be updated if Apple adds more symbols in future updates. For now, they
all seem to occupy the low end of Supplementary Private Use Area-B.

Note that these symbols only work when viewed in macOS, and won't work
in other platforms like Linux or Windows. They should probably not be
used in files that need to be shared outside of macOS.