macOS 11: SF symbols for preference pane / use the right toolbar styles
Use SF symbols (only for macOS 11+ / Big Sur) for preference pane's
toolbar to be consistent with rest of OS. Set toolbar style to be
"preference" as otherwise it defaults to "unified" which is not correct
for a preference pane.

Set toolbar style for main MacVim window to be "unified compact", as the
default "unified" is too larger and not useful for text editing. For
example, Xcode also uses this.