Update EdDSA key for Sparkle updater, as the old one was wrong
The EdDSA public key added in last release (snapshot-166) was wrong.
Simply re-generate a new pair. This is not fatal as Sparkle allows for
DSA-only updates for now, but fix it to have the right key so future
MacVim updates can be done via EdDSA. It won't work if someone is
updating directly from 166, but given enough time, that should not be very likely
as people woudl have updated to the newer versions.