Fix Travis CI to only build x86_64, and fix --with-macarchs flag
Apple Silicon build doesn't work in CI yet, and as such we don't want CI
to build it, as accidentaly Apple Silicon builds will mean an Apple
Silicon Mac prioritize that over Rosetta build which will crash (since
we are only building `MacVim` as universal, but the raw `Vim` process is
not). Also, add a CI smoketest step to make sure we only build x86_64.

Also, fix up the `--with-macarchs` flag for `configure`. For some
reason, Apple seems to have broken sed's word boundary parsing, so need
to work around it and use a more convoluted regex to do the word
splitting.