Update min OS requirement for snapshot-168 to macOS 10.14
The latest build broke 10.13 due to a build issue linking against newer
symbols. Simply disable this version for older versions of macOS for now
before snapshot-169 fixes this.

Also, include snapshot-166 in the appcast so people running macOS 10.13
or below will still be able to update to it successfully and Sparkle
will show that their MacVim is up to date instead of an error message.