Add matrix testing to GitHub Actions CI
GitHub Actions runners choices are relatively limited, so do matrix
testing on the two OS versions they support, and Xcode 11 to test
compatibility with Catalina (10.15) SDKs.

Remove the standard Vim's CI file, to make it less confusing and to
prevent GitHub Actions from running it accidentally (the UI has the
ability to disable certain workflows but it seems a little buggy at
times).

Also fixed up how terminal Vim was linking against Carbon and Cocoa
which are unnecessary. It only needs AppKit for communicating between
MMBackend and the GUI.

Also switch to using clang to match what we had in Travis CI.