Code clean up / robustification and additional logging for render
Clean up code per clang analyzer. Make sure to initialize structs
properly and clean up potential memory leaks.

When drawing wide characters, clean up the other cell to be empty just
to be clean.

When updating fonts for a MacVim window, if it's not the main window,
make sure to not update the shared font manager as it's previously a
minor bug where it would do that.

Add additional debug logging for renderer to help debug potential
CoreText renderer issues. See #1164.