Add automation to auto-close PR/issues if they are waiting for response
If a PR or issue has been tagged with "Response Needed" on GitHub, this
will auto-mark as stale in 45 days, and then close it in 7 days
afterwards. This allows us to simply mark an issue as "Response Needed"
and not need to come back after a while to close the issue which is time
consuming and easy to forget, leading to lots of stale open issues.

Don't turn this on for all issues. Only do this for those explicitly
tagged as needing response.